refactor: simplify sources traits, moved logic into each source
