+++ src/output/csv.rs
@@ -2,7 +2,7 @@ use std::fs::File;
 use std::io::Write;
 
 use crate::core::displayer::{Displayer, Result, default_iterations_filename};
-use crate::core::profiler::{Iteration, Phase};
+use crate::core::profiler::types::{Iteration, Phase};
 use crate::core::sensor::Sensor;
 use crate::util::file::{create_file_with_user_permissions, get_absolute_path};
