+++ src/output/json.rs
@@ -4,7 +4,7 @@ use std::io::Write;
 use serde_json::json;
 
 use crate::core::displayer::{Displayer, Result, default_iterations_filename};
-use crate::core::profiler::Iteration;
+use crate::core::profiler::types::Iteration;
 use crate::core::sensor::Sensor;
 use crate::util::file::{create_file_with_user_permissions, get_absolute_path};
